Added maven-antrun-plugin and updated paths for .app bundle
The maven-antrun-plugin was introduced to the pom.xml and necessary configurations were made to ensure the final jar file is properly copied to the correct directory within the .app bundle. Paths in several files (launch.sh, Info.plist, CommandLineController.java) were updated to optimally run the application whether or not it is from the .app bundle. Minor enhancements were made to the Info.plist for the bundle.