Implement image building from Dockerfile
Enabled uploading of Dockerfile in `DockerController` and added the capability to build images from Dockerfiles in `DockerManager`. Also, the Docker ping utility was moved to a more appropriate place in the same class to check if Docker is running or not. Now files can be uploaded and Docker images created from the uploaded Dockerfile, enhancing the functionality of the Docker management system.

Refactor UI layout and implement Dockerfile image build

The user interface has been improved by replacing the VBox with a TabPane in the DockerScene. The new layout consists of a "Predefined List" tab and a "Dockerfile" tab. Additionally, a new feature to build an image from a Dockerfile has been added to the DockerManager and integrated into the DockerController.