Refactor view handling and add Docker view
This commit streamlines the way the MenuBarController handles views by utilizing the passed-in view directly, rather than explicitly stating each case. It also includes Docker as a new view in ApplicationController and ApplicationBuilder. A minor change has also been done in the CLInteractor file where a to-do is added to abstract domain objects to a separate package.