Refactor MenuBarBuilder to improve UX
Updated the "Command Line Tools" item to be a standard MenuItem instead of a button for consistency in the 'Main Menu'. For the exit option, it has been updated from a menu item to a button to highlight its importance and functionality. This refactoring should make for a more consistent and user-friendly experience in the application.