Update DockerViewBuilder with Docker status checking
The DockerViewBuilder class was updated to include a 'Connect Docker' button and label indicating Docker's connection status. Modifications include the import and usage of new dependencies, addition of action handlers, methods for creating styled VBoxes and buttons, as well as refactoring the DockerViewBuilder's constructor to accept more arguments. The code for setting up the UI was updated accordingly.