Update DockerEngine to return detailed container info
The DockerEngine class has been updated to provide more detailed information when creating and running a container. Now, instead of returning a simple container ID, the `createAndRunContainer` method returns a `DockerContainer` object populated with information from the `InspectContainerResponse`. A call to check if Docker is running has also been added in the initial setup.